[TRE-3193]: Consolidate spec fixes into fix_openapi_spec.py
